refactor: [YSS-5] 패키지 정리
