--- yakssok/src/main/java/server/yakssok/domain/medication/AlarmSound.java
@@ -1,5 +1,0 @@
-package server.yakssok.domain.medication;
-
-
-public enum AlarmSound {
-}
