--- yakssok/src/main/java/server/yakssok/domain/medication/Cycle.java
@@ -1,6 +1,0 @@
-package server.yakssok.domain.medication;
-
-import jakarta.persistence.Entity;
-
-public enum Cycle {
-}
