--- yakssok/src/main/java/server/yakssok/domain/medication/MedicationStatus.java
@@ -1,8 +1,0 @@
-package server.yakssok.domain.medication;
-
-import jakarta.persistence.Entity;
-
-public enum MedicationStatus {
-	TAKING,
-	COMPLETED
-}
