+++ yakssok/src/main/java/server/yakssok/domain/medication/domain/entity/AlarmSound.java
@@ -1,0 +1,5 @@
+package server.yakssok.domain.medication.domain.entity;
+
+
+public enum AlarmSound {
+}
