+++ yakssok/src/main/java/server/yakssok/domain/medication/domain/entity/Cycle.java
@@ -1,0 +1,4 @@
+package server.yakssok.domain.medication.domain.entity;
+
+public enum Cycle {
+}
