+++ yakssok/src/main/java/server/yakssok/domain/medication/domain/entity/MedicationStatus.java
@@ -1,0 +1,6 @@
+package server.yakssok.domain.medication.domain.entity;
+
+public enum MedicationStatus {
+	TAKING,
+	COMPLETED
+}
