+++ yakssok/src/main/java/server/yakssok/domain/user/domain/entity/Provider.java
@@ -1,4 +1,4 @@
-package server.yakssok.domain.user;
+package server.yakssok.domain.user.domain.entity;
 
 import lombok.Getter;
